requirements.txt
pandas
numpy
xlrd
pyarrow
apache-airflow